Basic case for constructing classes works
closes #6
